refactor: share buildDynamicApiUrl util
