+++ src/utils/index.ts
@@ -5,3 +5,4 @@ export * from './sanitize.utils';
 export * from './safe-parsers.utils';
 export * from './validations';
 export * from './safe-local-storage';
+export * from './api-url';
